vim: symlink to nvim instead of alias or sh-script
Neovim recently [1] fixed their handling of v:progname and v:argv.
My neovim config has now been updated to use those to determine whether
to run with "vim" mode enabled or not.

To facilitate that mechanism to work neovim needs to be invoked as "vim"
which is most easily done by symlinking ~/.local/bin/vim to it, and then
invoking the symlink.

[1]: https://github.com/neovim/neovim/commit/0ce065a332cbd3874bd68a86fa9eda20c5467170