Added sway,foot,zathura configs
These are mostly copied from my previous (Void Linux) dotfiles.
Almost everything uses the Recursive Sans font, except for foot, which
now uses the MonoLisa font I bought :)

These, combined with the yambar config included in the last commit,
create a very minimaL tiling wayland experience. This is deliberately
EXTREMELY MINIMAL, as this is NOT designed to be run independently.

The intention behind this config is to have a minimal tiling wm
environment so that we can open multiple PDFs at the same time in
Zathura, and have all of them side-by-side in tabs. This is helpful for
doing pre-exam revision stuff.

This is designed to run ALONGSIDE the main GNOME session.
  1. from gdm, you log in to gnome, and do all of your stuff there.
  2. you switch to ttyN using alt-ctrl-fN
  3. you log in there using username & password
  4. you execute
  	$ dbus-run-session sway
  5. you now have a minimal sway session to study in.

> PLEASE NOTE:
Currently, sway DOES NOT LOCK the screen (we need swaylock for that)
So, please switch back to gnome whenever you wanna sleep/lock/lid-close.

P.S.
  If you are using a 1920x1080 screen, and have the correct fonts
  installed, then the terminal's dimensions should match the window
  dimensions EXACTLY.

  i.e. there should not be any space between the last line of the
  terminal, and the last line of the physical screen.

  Try running `man man`. The bottom-left corner should have a black text
  box (courtesy of the `less` pager) which should align perfectly with
  the physical corner of the screen :)

  Satisfying, isn't it?